pyproject.toml
[tool.poetry]
name = "telegram-weather-bot"
version = "0.1.0"
description = ""
authors = ["Patrick <pvilchez794@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_weather_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-telegram-bot = "^20.3"
pydantic = "^1.10.9"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
